% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Title}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{multiplier 1}

\item{b}{multiplier 2}
}
\description{
Title
}
\examples{
# Find the result of multiplication of 5 with 2
# multiply(5,2)
}
